Initの関数を別ファイルにまとめた
